+++ src/js/module/dashboard/ListIssuesAssigned.js
@@ -96,7 +96,7 @@ class ListIssuesAssigned extends React.Component {
                 <div className="pt-4">
                     <PanelIssues
                         title="No Priority"
-                        extraClass="none"
+                        extraClass="no-priority"
                         hideOnEmpty
                         // eslint-disable-next-line max-len
                         data={_.pick(this.props.issues, issue => _.intersection(_.map(issue.labels, label => label.name), ['Hourly', 'Daily', 'Weekly', 'Monthly']).length === 0)}